+++ library/src/main/java/com/getstream/sdk/chat/adapter/AttachmentViewHolderFactory.kt
@@ -22,10 +22,8 @@ open class AttachmentViewHolderFactory {
     }
 
     lateinit var listenerContainer: ListenerContainer
-        @JvmName("setListenerContainerInternal")
         internal set
     lateinit var bubbleHelper: MessageListView.BubbleHelper
-        @JvmName("setBubbleHelperInternal")
         internal set
 
     open fun getAttachmentViewType(