+++ library/src/test/java/androidx/arch/core/executor/testing/InstantExecutorExtension.kt
@@ -10,13 +10,15 @@ class InstantExecutorExtension : BeforeEachCallback, AfterEachCallback {
 
     override fun beforeEach(context: ExtensionContext?) {
         ArchTaskExecutor.getInstance()
-            .setDelegate(object : TaskExecutor() {
-                override fun executeOnDiskIO(runnable: Runnable) = runnable.run()
+            .setDelegate(
+                object : TaskExecutor() {
+                    override fun executeOnDiskIO(runnable: Runnable) = runnable.run()
 
-                override fun postToMainThread(runnable: Runnable) = runnable.run()
+                    override fun postToMainThread(runnable: Runnable) = runnable.run()
 
-                override fun isMainThread(): Boolean = true
-            })
+                    override fun isMainThread(): Boolean = true
+                }
+            )
     }
 
     override fun afterEach(context: ExtensionContext?) {