+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Replace Glide with Coil in AttachmentViewHolderMedia (Fix GIFs loading issues)
 - Update Offline Support to the [last version](. See changes: )https://github.com/GetStream/stream-chat-android-livedata/releases/tag/0.8.6)
 - Allow setting custom `NotificationHandler` in `Chat.Builder`
+- `MessageListView.BubbleHelper`'s methods now have nullability annotations, and use primitive `boolean` values as parameters
 
 # Oct 14th, 2020 - 4.3.0-beta-6
 - Update to Kotlin 1.4.10